+++ packages/core/src/utils/SafeLoopArray.ts
